refactor: split i18n locale modules
