+++ src/i18n/locales/en.ts
@@ -1,0 +1,317 @@
+import type { TranslationSet } from '../types'
+
+export const en: TranslationSet = {
+  appTitle: 'Subscription Payoff',
+  appSubtitle: 'Codex Pacer',
+  nav: {
+    overview: 'Overview',
+    conversations: 'Conversations',
+  },
+  actions: {
+    rescanNow: 'Rescan now',
+    refreshPricing: 'Refresh pricing',
+    settings: 'Settings',
+    close: 'Close',
+    cancel: 'Cancel',
+    saveSettings: 'Save settings',
+    saving: 'Saving…',
+  },
+  common: {
+    autoScan: 'Auto scan',
+    disabled: 'Disabled',
+    everyMinutes: (minutes) => `Every ${minutes} min`,
+    quotaWindow: 'Quota window',
+    earlier: 'Earlier',
+    newer: 'Newer',
+    start: 'Start',
+    end: 'End',
+    reset: 'Reset',
+    updated: 'Updated',
+    tokens: 'Tokens',
+    sessions: 'sessions',
+    session: 'Session',
+    fast: 'Fast',
+    noData: 'No data',
+    searchTitleOrSession: 'Search title or session id',
+    customRangeStartDate: 'Custom range start date',
+    customRangeEndDate: 'Custom range end date',
+  },
+  status: {
+    waitingForFirstScan: 'Waiting for first scan…',
+    fetchingLiveQuotaWindow: 'Fetching live quota window…',
+    scannedFiles: (files, sessions) => `Scanned ${files} files, refreshed ${sessions} sessions.`,
+    backgroundScanAlreadyRunning: 'Background scan is already running. Refreshing local snapshots.',
+    failedToLoad: (bucketLabel, error) => `Failed to load ${bucketLabel}: ${error}`,
+    pricingRefreshed: 'Pricing catalog refreshed from OpenAI Standard short-context pricing.',
+    settingsSaved: 'Settings saved.',
+    subscriptionRecordSaved: 'Subscription record saved.',
+    subscriptionRecordDeleted: 'Subscription record deleted.',
+    waitingLiveQuota: 'Waiting for live quota',
+  },
+  buckets: {
+    five_hour: '5h',
+    day: 'Day',
+    seven_day: '7d',
+    week: 'Week',
+    subscription_month: 'Billing month',
+    month: 'Month',
+    year: 'Year',
+    custom: 'Custom',
+    total: 'Total',
+  },
+  bucketDescriptions: {
+    subscriptionMonth: (day) => `Starts on day ${day} each month`,
+    currentFiveHourWindow: 'Current 5-hour quota window',
+    historicalFiveHourWindow: 'Historical 5-hour quota window',
+    currentSevenDayWindow: 'Current 7-day quota window',
+    historicalSevenDayWindow: 'Historical 7-day quota window',
+  },
+  metrics: {
+    apiValue: 'API value (Std short ctx)',
+    wastedQuota: 'Wasted quota',
+    remainingQuota: 'Remaining quota',
+    peakUsed: (value) => `Peak used ${value}`,
+    used: (value) => `Used ${value}`,
+    windowStart: 'Window start',
+    remainingTime: 'Time remaining',
+    totalTokens: 'Total tokens',
+    conversationCount: 'Conversations',
+    subscriptionCost: 'Subscription cost',
+    payoffRatio: 'Payoff ratio',
+    planPerMonth: (planType, price) => `${planType} · ${price}/mo`,
+    subscriptionLedgerNote: (count) => `${count} subscription records`,
+    noSubscriptionRecords: 'No subscription records',
+    monthlyFeeShare: 'Share of subscription ledger',
+  },
+  overview: {
+    distribution: 'Distribution',
+    modelShare: 'Model share',
+    costStructure: 'Cost structure',
+    sourceShare: 'Source share',
+  },
+  conversationList: {
+    eyebrow: 'Conversation list',
+    title: 'Root sessions',
+    shown: (count) => `${count} shown`,
+    empty: 'No conversations matched the current filter.',
+  },
+  detail: {
+    eyebrow: 'Conversation detail',
+    modelBreakdown: 'Model breakdown',
+    conversationModelBreakdown: 'Conversation model breakdown',
+    conversationCostBreakdown: 'Conversation cost structure',
+    conversationSourceBreakdown: 'Conversation source breakdown',
+    turnTimelineEyebrow: 'Turn timeline',
+    turnUsage: 'Turn usage',
+    latestTurns: (count) => `Latest ${count} turns`,
+    emptyTurns: 'No turn-level source data is available for this conversation.',
+    emptySelection: 'Select a conversation to inspect its detailed ledger.',
+    untitledTurn: 'Untitled turn',
+    unknownModel: 'Unknown model',
+    subagent: 'Subagent',
+    mainSession: 'Main session',
+    sessionLabel: (sessionId) => `Session ${sessionId.slice(0, 8)}`,
+  },
+  charts: {
+    trendEyebrow: 'Trend',
+    valueAndTokens: 'Value and tokens',
+    valueVsTokens: 'Value / Tokens',
+    apiValue: 'API value',
+    tokens: 'Tokens',
+    noShareData: 'No share data in this window.',
+    valueUnavailableTokenFallback: 'API value is unavailable, showing token share instead.',
+    dimensionControlLabel: 'Distribution dimension',
+    metricControlLabel: 'Distribution metric',
+    byModel: 'By model',
+    byStructure: 'By structure',
+    bySource: 'By source',
+    byValue: 'By value',
+    byTokens: 'By tokens',
+    liveQuotaEyebrow: 'Live quota',
+    fiveHourQuotaTrend: '5-hour quota trend',
+    sevenDayQuotaTrend: '7-day quota trend',
+    historicalQuotaNote: 'Full historical window / remaining quota / cumulative value',
+    currentQuotaNote: 'Current window start to reset / remaining quota / cumulative value',
+    noLiveQuotaHistory: 'No live quota history is available for the current window yet.',
+    remaining: 'Remaining',
+    cumulativeValue: 'Cumulative value',
+    windowValue: 'Window value',
+  },
+  sources: {
+    label: 'Sources',
+    remoteSources: 'Remote sources',
+    chooseSources: 'Choose sources for this view',
+    noneSelected: 'None selected',
+    listSeparator: ', ',
+    local: 'Local',
+    localCodexHome: 'Local ~/.codex',
+    cached: 'Cached',
+    added: 'Added',
+    failed: 'Failed',
+    downloading: 'Downloading',
+    notDownloaded: 'Not downloaded',
+    addSsh: 'Add SSH',
+    updateAll: 'Update all',
+    noRemoteSources: 'No remote sources yet.',
+    update: 'Update',
+    updating: 'Updating',
+    deleteSource: 'Delete source',
+    deleteSourceLabel: (label) => `Delete ${label}`,
+    deleting: 'Deleting…',
+    deleteServer: 'Delete server',
+    deleteServerTitle: (label) => `Delete ${label}?`,
+    deleteServerDescription: 'This removes the SSH source, its local cache, and imported stats. The remote ~/.codex is not changed.',
+    confirmDelete: 'Delete',
+    addCodexServer: 'Add Codex server',
+    refreshSshList: 'Refresh SSH list',
+    filteredHostsNote: 'GitHub/GitLab-style hosts are filtered automatically.',
+    noSshServersDiscovered: 'No SSH servers discovered.',
+    add: 'Add',
+    keepOneSelected: 'Keep at least one source selected.',
+    addedSource: (label) => `Added ${label}`,
+    downloadingRemoteCache: 'Downloading remote cache…',
+    downloadedAndImported: (files) => `Downloaded and imported ${files} files`,
+    noSshServersAdded: 'No SSH servers added yet.',
+    deletingSource: (label) => `Deleting ${label}…`,
+    deletedSource: (label) => `Deleted ${label}`,
+    maintenanceActions: 'Maintenance actions',
+  },
+  popup: {
+    title: 'Tray',
+    updated: (value) => `Updated · ${value}`,
+    loading: 'Loading tray snapshot…',
+    failedTitle: 'Failed to load tray snapshot',
+    resetIn: (value) => `Resets in ${value}`,
+    speedTitle: '7D pace',
+    speedHint: '100% = quota and time are aligned; below 100% is faster, above 100% is slower',
+    sevenDayUsageChart: '7-day usage line chart',
+    chartLegendRemaining: 'Blue: quota left',
+    chartLegendReference: 'Dashed: time baseline',
+    chartLegendCurrent: 'Dot: current',
+    chartValueBadge: 'API value',
+    speedStatus: {
+      fast: 'Slow down',
+      healthy: 'On pace',
+      slow: 'Push harder',
+    },
+    resetTimeline5h: '5h reset',
+    resetTimeline7d: '7d reset',
+    modules: {
+      apiValue: 'API',
+      tokenCount: 'Tokens',
+      scanFreshness: 'Last scan',
+      liveQuotaFreshness: 'Quota age',
+      payoffRatio: 'Payoff',
+      conversationCount: 'Chats',
+    },
+    actions: {
+      openDashboard: 'Dashboard',
+      refresh: 'Refresh',
+      settings: 'Settings',
+    },
+  },
+  settings: {
+    appSettings: 'App settings',
+    syncAndSubscriptionProfile: 'Sync and subscription profile',
+    sections: {
+      language: {
+        eyebrow: 'Language',
+        title: 'Interface language',
+        description: 'Switch the app language. The current structure supports Chinese and English now while keeping room for more languages later.',
+        label: 'Display language',
+        note: 'This preference is stored locally and does not change synced settings.',
+      },
+      sync: {
+        eyebrow: 'Sync',
+        title: 'Scan and data sources',
+        description: 'Control the Codex data directory, automatic scan cadence, and live quota refresh interval.',
+        codexHome: 'Codex home',
+        codexHomePlaceholder: 'Defaults to CODEX_HOME or ~/.codex',
+        autoScanEnabled: 'Auto scan enabled',
+        autoScanEnabledNote: 'When disabled, only manual scans are kept and periodic refresh stops.',
+        autoScanIntervalMinutes: 'Auto scan interval (minutes)',
+        liveQuotaRefreshIntervalSeconds: 'Live quota refresh interval (minutes)',
+        liveQuotaRefreshNote: 'Separately controls active refresh and history persistence for `5h / 7d` live quota snapshots.',
+      },
+      menuBar: {
+        eyebrow: 'Tray',
+        title: 'Menu bar / system tray display',
+        description: 'Manage the logo, amount, live quota metric, and source used in the tray.',
+        hideDockIcon: 'Hide Dock icon while menu bar is visible',
+        hideDockIconNote: 'Uses the menu bar item as the app entry point. If the menu bar item is hidden, the Dock icon stays visible.',
+        showLogo: 'Show logo',
+        showLogoNote: 'When disabled, only text stays visible. If every text metric is also off, the tray item is hidden.',
+        showApiValue: 'Show API value',
+        showApiValueNote: 'You can turn this off and still keep the logo or live quota metric visible.',
+        showLiveQuotaMetric: 'Show live quota metric',
+        showLiveQuotaMetricNote: 'Adds one live quota metric to the right side of the amount.',
+        range: 'Tray range',
+        rangeNote: 'Choose the aggregation window for the tray amount. `5h / 7d` uses the current live rate limit window.',
+        liveMetric: 'Tray live metric',
+        liveMetricRemainingPercent: 'Remaining percent',
+        liveMetricSuggestedUsageSpeed: 'Suggested usage speed %',
+        liveMetricNote: '`Suggested usage speed % = remaining quota ÷ remaining time share × 100%`. `100%` means the pace is balanced; lower means you are burning quota faster, higher means you can be more aggressive.',
+        quotaSource: 'Tray quota source',
+        quotaSourceNote: 'Choose whether the tray uses the `5h` or `7d` quota window as the single metric source.',
+        speedEmojiSection: 'Speed status emoji',
+        speedShowEmoji: 'Show speed-status emoji',
+        speedShowEmojiNote: 'When disabled, the tray shows only the percentage without status icons.',
+        speedFastThreshold: 'Fast threshold (%)',
+        speedSlowThreshold: 'Slow threshold (%)',
+        speedThresholdNote: 'Below the fast threshold shows the fast-usage emoji; above the slow threshold shows the slow-usage emoji; the middle range shows the healthy emoji.',
+        speedHealthyEmoji: 'Healthy emoji',
+        speedFastEmoji: 'Too-fast emoji',
+        speedSlowEmoji: 'Too-slow emoji',
+        speedEmojiNote: 'Any short emoji or symbol works, but shorter values keep the tray cleaner.',
+        popupEnabled: 'Enable tray popup',
+        popupEnabledNote: 'Left-clicking the tray item opens a compact quota popup with the most important metrics.',
+        popupModules: 'Popup secondary modules',
+        popupModulesNote: 'Core quota cards always stay visible; use this list to control extra cards and their order.',
+        popupShowResetTimeline: 'Show reset timeline row',
+        popupShowActions: 'Show footer action buttons',
+        moveUp: 'Move up',
+        moveDown: 'Move down',
+      },
+      subscription: {
+        eyebrow: 'Subscription',
+        title: 'Subscription framing',
+        description: 'These parameters directly affect payoff, billing-month slicing, and related cost calculations.',
+        planType: 'Plan type',
+        currency: 'Currency',
+        currencyNote: 'Cost and payoff metrics currently support USD only. If you previously entered another currency, update the monthly price to its USD amount.',
+        monthlyPrice: 'Monthly subscription price',
+        billingAnchorDay: 'Subscription month starts on',
+        billingAnchorDayNote: 'The billing month is split by this day. For example, `23` means each window runs from the 23rd to the 22nd of the next month.',
+        accountStatus: 'Current Codex login',
+        accountUnavailable: 'Unavailable',
+        accountRequiresLogin: 'OpenAI login required',
+        accountApiKey: 'API key mode',
+        accountUnknown: 'Unknown account',
+        planPlus: 'ChatGPT Plus',
+        planPro5: 'ChatGPT Pro 5x',
+        planPro10: 'ChatGPT Pro 10x',
+        amountUsd: 'Amount (USD)',
+        serviceStart: 'Service start',
+        serviceEnd: 'Service end',
+        addRecordTitle: 'Ledger record',
+        editRecordTitle: 'Edit record',
+        addRecord: 'Add subscription record',
+        updateRecord: 'Update subscription record',
+        saveRecord: 'Save record',
+        editRecord: 'Edit',
+        removeRecord: 'Delete',
+        cancelEditRecord: 'Cancel edit',
+        emptyRecords: 'No subscription ledger records yet. Add records to prorate payoff by service period overlap.',
+        accountEmail: 'Account email / note',
+        accountEmailPlaceholder: 'Optional note or email',
+      },
+      liveQuota: {
+        eyebrow: 'Live Quota',
+        title: 'Current quota windows',
+        description: 'The `5h / 7d` overview views and this section share the same live quota snapshot and refresh cadence.',
+        remaining: (value) => `Remaining ${value}`,
+        timeLeft: (value) => `${value} left`,
+      },
+    },
+  },
+}
