+++ src/i18n/locales/zhCN.ts
@@ -1,0 +1,317 @@
+import type { TranslationSet } from '../types'
+
+export const zhCN: TranslationSet = {
+  appTitle: '订阅回本盘',
+  appSubtitle: 'Codex Pacer',
+  nav: {
+    overview: '总览',
+    conversations: '对话',
+  },
+  actions: {
+    rescanNow: '立即扫描',
+    refreshPricing: '刷新定价',
+    settings: '设置',
+    close: '关闭',
+    cancel: '取消',
+    saveSettings: '保存设置',
+    saving: '保存中…',
+  },
+  common: {
+    autoScan: '自动扫描',
+    disabled: '已关闭',
+    everyMinutes: (minutes) => `每 ${minutes} 分钟`,
+    quotaWindow: '额度窗口',
+    earlier: '更早',
+    newer: '较新',
+    start: '开始',
+    end: '结束',
+    reset: '重置',
+    updated: '更新于',
+    tokens: 'Tokens',
+    sessions: '会话',
+    session: '会话',
+    fast: 'Fast',
+    noData: '暂无数据',
+    searchTitleOrSession: '搜索标题或会话 ID',
+    customRangeStartDate: '自定义范围开始日期',
+    customRangeEndDate: '自定义范围结束日期',
+  },
+  status: {
+    waitingForFirstScan: '等待首次扫描…',
+    fetchingLiveQuotaWindow: '正在获取 live quota 窗口…',
+    scannedFiles: (files, sessions) => `已扫描 ${files} 个文件，刷新 ${sessions} 个会话。`,
+    backgroundScanAlreadyRunning: '后台扫描已在进行中，正在刷新本地快照。',
+    failedToLoad: (bucketLabel, error) => `加载${bucketLabel}失败：${error}`,
+    pricingRefreshed: '已按 OpenAI 标准短上下文定价刷新目录。',
+    settingsSaved: '设置已保存。',
+    subscriptionRecordSaved: '订阅记录已保存。',
+    subscriptionRecordDeleted: '订阅记录已删除。',
+    waitingLiveQuota: '等待 live quota',
+  },
+  buckets: {
+    five_hour: '5小时',
+    day: '日',
+    seven_day: '7天',
+    week: '周',
+    subscription_month: '订阅月',
+    month: '月',
+    year: '年',
+    custom: '自定义',
+    total: '总计',
+  },
+  bucketDescriptions: {
+    subscriptionMonth: (day) => `每月${day}号起`,
+    currentFiveHourWindow: '当前 5 小时额度窗口',
+    historicalFiveHourWindow: '历史 5 小时额度窗口',
+    currentSevenDayWindow: '当前 7 天额度窗口',
+    historicalSevenDayWindow: '历史 7 天额度窗口',
+  },
+  metrics: {
+    apiValue: 'API 等价价值（标准短上下文）',
+    wastedQuota: '浪费额度',
+    remainingQuota: '剩余额度',
+    peakUsed: (value) => `峰值已用 ${value}`,
+    used: (value) => `已用 ${value}`,
+    windowStart: '时间段开始',
+    remainingTime: '剩余时间',
+    totalTokens: 'Tokens 总量',
+    conversationCount: '对话数',
+    subscriptionCost: '订阅费',
+    payoffRatio: '回本比例',
+    planPerMonth: (planType, price) => `${planType} · ${price}/月`,
+    subscriptionLedgerNote: (count) => `${count} 条订阅记录`,
+    noSubscriptionRecords: '未添加订阅记录',
+    monthlyFeeShare: '订阅账本占比',
+  },
+  overview: {
+    distribution: '分布',
+    modelShare: '模型占比',
+    costStructure: '成本结构',
+    sourceShare: '数据源占比',
+  },
+  conversationList: {
+    eyebrow: '对话列表',
+    title: '根会话',
+    shown: (count) => `显示 ${count} 条`,
+    empty: '当前筛选条件下没有匹配对话。',
+  },
+  detail: {
+    eyebrow: '对话详情',
+    modelBreakdown: '模型分布',
+    conversationModelBreakdown: '对话模型分布',
+    conversationCostBreakdown: '对话成本结构',
+    conversationSourceBreakdown: '对话数据源分布',
+    turnTimelineEyebrow: 'Turn 时间线',
+    turnUsage: 'Turn 用量',
+    latestTurns: (count) => `最近 ${count} 个 turn`,
+    emptyTurns: '该对话暂无 turn 级来源数据。',
+    emptySelection: '选择一个对话以查看详细账本。',
+    untitledTurn: '未命名 turn',
+    unknownModel: '未知模型',
+    subagent: '子代理',
+    mainSession: '主会话',
+    sessionLabel: (sessionId) => `会话 ${sessionId.slice(0, 8)}`,
+  },
+  charts: {
+    trendEyebrow: '趋势',
+    valueAndTokens: '价值与 Tokens',
+    valueVsTokens: '价值 / Tokens',
+    apiValue: 'API 价值',
+    tokens: 'Tokens',
+    noShareData: '当前窗口没有可用占比数据。',
+    valueUnavailableTokenFallback: '暂无可用 API 价值，已按 tokens 显示占比。',
+    dimensionControlLabel: '分布维度',
+    metricControlLabel: '分布指标',
+    byModel: '按模型',
+    byStructure: '按结构',
+    bySource: '按数据源',
+    byValue: '按价值',
+    byTokens: '按 tokens',
+    liveQuotaEyebrow: 'Live quota',
+    fiveHourQuotaTrend: '5 小时额度轨迹',
+    sevenDayQuotaTrend: '7 天额度轨迹',
+    historicalQuotaNote: '完整历史窗口 / 剩余额度 / 累积价值',
+    currentQuotaNote: '当前窗口起点到重置点 / 剩余额度 / 累积价值',
+    noLiveQuotaHistory: '当前窗口还没有可用的 live quota 历史记录。',
+    remaining: '剩余',
+    cumulativeValue: '累积价值',
+    windowValue: '窗口价值',
+  },
+  sources: {
+    label: '数据源',
+    remoteSources: '远程数据源',
+    chooseSources: '选择要纳入统计的数据源',
+    noneSelected: '未选择',
+    listSeparator: '、',
+    local: '本机',
+    localCodexHome: '本机 ~/.codex',
+    cached: '已缓存',
+    added: '已添加',
+    failed: '失败',
+    downloading: '下载中',
+    notDownloaded: '未下载',
+    addSsh: '添加 SSH',
+    updateAll: '更新全部',
+    noRemoteSources: '还没有远程数据源。',
+    update: '更新',
+    updating: '更新中',
+    deleteSource: '删除数据源',
+    deleteSourceLabel: (label) => `删除 ${label}`,
+    deleting: '删除中…',
+    deleteServer: '删除服务器',
+    deleteServerTitle: (label) => `删除 ${label}？`,
+    deleteServerDescription: '这会移除这个 SSH 数据源、本地缓存和已导入的统计数据。不会影响远程服务器上的 ~/.codex。',
+    confirmDelete: '确认删除',
+    addCodexServer: '添加 Codex 服务器',
+    refreshSshList: '刷新 SSH 列表',
+    filteredHostsNote: '自动过滤 GitHub / GitLab 等代码托管 Host。',
+    noSshServersDiscovered: '没有发现可添加的 SSH 服务器。',
+    add: '添加',
+    keepOneSelected: '至少保留一个数据源。',
+    addedSource: (label) => `已添加 ${label}`,
+    downloadingRemoteCache: '正在下载远程缓存…',
+    downloadedAndImported: (files) => `已下载并导入 ${files} 个文件`,
+    noSshServersAdded: '还没有添加 SSH 服务器。',
+    deletingSource: (label) => `正在删除 ${label}…`,
+    deletedSource: (label) => `已删除 ${label}`,
+    maintenanceActions: '维护操作',
+  },
+  popup: {
+    title: '托盘',
+    updated: (value) => `更新 · ${value}`,
+    loading: '正在加载托盘概览…',
+    failedTitle: '托盘概览加载失败',
+    resetIn: (value) => `${value}后重置`,
+    speedTitle: '7天节奏',
+    speedHint: '100% = 额度消耗与时间同步；低于 100% 偏快，高于 100% 偏慢',
+    sevenDayUsageChart: '7 天使用情况折线图',
+    chartLegendRemaining: '剩余额度',
+    chartLegendReference: '时间基准',
+    chartLegendCurrent: '当前',
+    chartValueBadge: 'API 价值',
+    speedStatus: {
+      fast: '放慢',
+      healthy: '正常',
+      slow: '可加速',
+    },
+    resetTimeline5h: '5h reset',
+    resetTimeline7d: '7d reset',
+    modules: {
+      apiValue: 'API',
+      tokenCount: 'Tokens',
+      scanFreshness: '上次扫描',
+      liveQuotaFreshness: '额度快照',
+      payoffRatio: '回本',
+      conversationCount: '对话',
+    },
+    actions: {
+      openDashboard: '主窗口',
+      refresh: '刷新',
+      settings: '设置',
+    },
+  },
+  settings: {
+    appSettings: '应用设置',
+    syncAndSubscriptionProfile: '同步与订阅配置',
+    sections: {
+      language: {
+        eyebrow: '语言',
+        title: '界面语言',
+        description: '切换应用展示语言，当前实现支持中英双语，并为后续新增语言保留扩展位。',
+        label: '显示语言',
+        note: '语言偏好保存在本地设备，不影响已有同步配置。',
+      },
+      sync: {
+        eyebrow: '同步',
+        title: '扫描与数据源',
+        description: '控制 Codex 数据目录、自动扫描频率，以及 live quota 的刷新周期。',
+        codexHome: 'Codex home',
+        codexHomePlaceholder: '默认使用 CODEX_HOME 或 ~/.codex',
+        autoScanEnabled: '启用自动扫描',
+        autoScanEnabledNote: '关闭后仅保留手动扫描，不再按周期自动刷新数据。',
+        autoScanIntervalMinutes: '自动扫描间隔（分钟）',
+        liveQuotaRefreshIntervalSeconds: 'Live quota 刷新间隔（分钟）',
+        liveQuotaRefreshNote: '独立控制 `5小时 / 7天` live quota 的主动刷新与历史持久化频率。',
+      },
+      menuBar: {
+        eyebrow: '托盘',
+        title: '菜单栏 / 系统托盘显示',
+        description: '统一管理托盘里的 logo、金额、live quota 指标和指标来源。',
+        hideDockIcon: '菜单栏可见时隐藏程序坞图标',
+        hideDockIconNote: '启用后应用以菜单栏作为入口；若菜单栏项目被隐藏，程序坞图标会保持显示。',
+        showLogo: '显示 logo',
+        showLogoNote: '关闭后仅保留文字内容；若同时关闭所有文字指标，托盘项目会隐藏。',
+        showApiValue: '显示 API 价值',
+        showApiValueNote: '关闭后仍可单独保留 logo 或 live quota 指标。',
+        showLiveQuotaMetric: '显示 live quota 指标',
+        showLiveQuotaMetricNote: '在金额右侧附加显示一个 live quota 指标。',
+        range: '托盘统计范围',
+        rangeNote: '选择托盘金额对应的统计尺度。`5小时 / 7天` 会使用 Codex live rate limits 当前窗口。',
+        liveMetric: '托盘 live 指标',
+        liveMetricRemainingPercent: '剩余百分比',
+        liveMetricSuggestedUsageSpeed: '建议使用速度百分比',
+        liveMetricNote: '`建议使用速度百分比 = 剩余额度 ÷ 剩余时间占比 × 100%`。`100%` 表示节奏健康；越低说明用得越快，越高说明还能更积极使用。',
+        quotaSource: '托盘额度来源',
+        quotaSourceNote: '指定托盘右侧使用 `5小时` 或 `7天` 额度窗口作为指标来源，只显示一个。',
+        speedEmojiSection: '速度状态 Emoji',
+        speedShowEmoji: '显示速度状态 emoji',
+        speedShowEmojiNote: '关闭后托盘只显示百分比，不显示健康/过快/过慢状态图标。',
+        speedFastThreshold: '过快阈值（%）',
+        speedSlowThreshold: '过慢阈值（%）',
+        speedThresholdNote: '低于“过快阈值”显示过快 emoji；高于“过慢阈值”显示过慢 emoji；中间区间显示健康 emoji。',
+        speedHealthyEmoji: '健康 emoji',
+        speedFastEmoji: '过快 emoji',
+        speedSlowEmoji: '过慢 emoji',
+        speedEmojiNote: '可输入任意短 emoji 或符号，建议保持简短以避免托盘拥挤。',
+        popupEnabled: '启用托盘弹窗',
+        popupEnabledNote: '左键点击托盘项目时，弹出一个紧凑面板展示额度和关键信息。',
+        popupModules: '弹窗附加模块',
+        popupModulesNote: '核心额度卡片始终显示；这里可控制附加信息卡片的显示与顺序。',
+        popupShowResetTimeline: '显示 reset 时间条',
+        popupShowActions: '显示底部操作按钮',
+        moveUp: '上移',
+        moveDown: '下移',
+      },
+      subscription: {
+        eyebrow: '订阅',
+        title: '订阅口径',
+        description: '这些参数会直接影响回本、订阅月和相关成本计算。',
+        planType: '套餐类型',
+        currency: '货币',
+        currencyNote: '当前金额与回本计算只支持 USD。若你原来填过其他币种，请把月费改成对应的美元金额。',
+        monthlyPrice: '月订阅价格',
+        billingAnchorDay: '订阅月起始日',
+        billingAnchorDayNote: '订阅月会按这个日期切分。例如设置为 23，则统计范围是每月 23 号到次月 22 号。',
+        accountStatus: '当前 Codex 登录',
+        accountUnavailable: '不可用',
+        accountRequiresLogin: '需要登录 OpenAI',
+        accountApiKey: 'API key 模式',
+        accountUnknown: '未知账号',
+        planPlus: 'ChatGPT Plus',
+        planPro5: 'ChatGPT Pro 5x',
+        planPro10: 'ChatGPT Pro 10x',
+        amountUsd: '金额（USD）',
+        serviceStart: '服务开始',
+        serviceEnd: '服务结束',
+        addRecordTitle: '账本记录',
+        editRecordTitle: '编辑记录',
+        addRecord: '添加订阅记录',
+        updateRecord: '更新订阅记录',
+        saveRecord: '保存记录',
+        editRecord: '编辑',
+        removeRecord: '删除',
+        cancelEditRecord: '取消编辑',
+        emptyRecords: '还没有订阅账本记录。添加记录后，回本会按服务期重叠比例计算。',
+        accountEmail: '账号邮箱/备注',
+        accountEmailPlaceholder: '可选备注或邮箱',
+      },
+      liveQuota: {
+        eyebrow: 'Live Quota',
+        title: '当前额度窗口',
+        description: '总览中的 `5小时 / 7天` 视图和这里使用同一份 live quota 快照与刷新周期。',
+        remaining: (value) => `剩余 ${value}`,
+        timeLeft: (value) => `还剩 ${value}`,
+      },
+    },
+  },
+}
